requirements.txt
fastapi[all]
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
alembic
passlib[bcrypt]
python-jose[cryptography]